refactored tests so the implemented algorithm, interview questions and tests are all in one file
